requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
mistralai>=0.0.12
python-dotenv>=1.0.0
pydantic>=2.0.0
